requirements.txt
python-dateutil
ipython
ipython-genutils
wheel
certifi
chardet
click
cython
filterpy
h5py
ruamel.yaml>=0.15.0
intel-openmp
imgaug==0.4.0
numba==0.51.1
matplotlib==3.1.3
networkx
numpy~=1.17.3
opencv-python-headless~=3.4.9.33
pandas>=1.0.1
patsy
pyyaml==5.4.1
setuptools
scikit-image>=0.17
scikit-learn
scipy>=1.4
six
statsmodels>=0.11
tables
tensorpack==0.9.8
tqdm
moviepy<=1.0.1
bayesian-optimization
wxPython
easydict
tensorflow==1.15.0
opencv-python==4.5.1.48